Separate code in different files and fix some old documentation
