+++ examples/starknet.rs
@@ -2,11 +2,12 @@
 
 use cairo_felt::Felt252;
 use cairo_lang_compiler::CompilerConfig;
+use cairo_native::context::NativeContext;
+use cairo_native::executor::NativeExecutor;
 use cairo_native::{
     metadata::syscall_handler::SyscallHandlerMeta,
     starknet::{BlockInfo, ExecutionInfo, StarkNetSyscallHandler, SyscallResult, TxInfo, U256},
     utils::find_function_id,
-    NativeContext, NativeExecutor,
 };
 use serde_json::json;
 use std::{io, path::Path};
